requirements.txt
-i https://pypi.org/simple
aiosqlite==0.17.0 ; python_version >= '3.6'
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==4.3.0 ; python_version >= '3.8'
async-lru==2.0.4
async-timeout==4.0.3 ; python_full_version < '3.12.0'
asyncpg==0.29.0
certifi==2024.2.2 ; python_version >= '3.6'
click==8.1.7 ; python_version >= '3.7'
colorlog==6.8.2
fastapi==0.110.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==1.0.4 ; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0
idna==3.6 ; python_version >= '3.5'
iso8601==1.1.0 ; python_version < '4.0' and python_full_version >= '3.6.2'
newrelic==9.7.0
numpy==1.26.4 ; python_version == '3.11'
pandas==2.2.1
psycopg==3.1.18
pydantic==2.6.2 ; python_version >= '3.8'
pydantic-core==2.16.3 ; python_version >= '3.8'
pydantic-settings==2.2.1
pypika-tortoise==0.1.6 ; python_version >= '3.7' and python_version < '4.0'
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.36.3 ; python_version >= '3.8'
tortoise-orm==0.20.0
typing-extensions==4.9.0 ; python_version >= '3.8'
tzdata==2024.1 ; python_version >= '2'
uvicorn[standard]==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
